# docs/api.yaml
openapi: 3.1.0
info:
  title: Tadpoles API Starter
  version: 1.0.0
  description: >-
    Starter OpenAPI contract for the Tadpoles Spring Boot application.
    The API assumes JWT authentication, Spring Security authorization,
    and role-aware responses for Admin, Auditor, Analyst, Advisor, and Client users.
servers:
  - url: /api
    description: Application API root
...
security:
  - bearerAuth: []
paths:
...
  /users/me:
    get:
      tags: [Users]
      summary: Get the currently authenticated user profile
      operationId: getCurrentUser
      x-required-roles: [ADMIN, AUDITOR, ANALYST, ADVISOR, CLIENT]
      responses:
        '200':
          description: Current user profile
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserProfile'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '429':
          $ref: '#/components/responses/TooManyRequests'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /users:
    get:
      tags: [Users]
      summary: List users for administration
      operationId: listUsers
      x-required-roles: [ADMIN]
      responses:
        '200':
          description: Paginated user list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserListResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '429':
          $ref: '#/components/responses/TooManyRequests'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /users/{userId}:
    parameters:
      - $ref: '#/components/parameters/UserId'
    get:
      tags: [Users]
      summary: Get a user by ID
      operationId: getUserById
      x-required-roles: [ADMIN]
      responses:
        '200':
          description: User record
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserProfile'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '429':
          $ref: '#/components/responses/TooManyRequests'
        '500':
          $ref: '#/components/responses/InternalServerError'
    patch:
      tags: [Users]
      summary: Update a user profile or status
      operationId: updateUser
      x-required-roles: [ADMIN]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateUserRequest'
      responses:
        '200':
          description: User updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserProfile'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '429':
          $ref: '#/components/responses/TooManyRequests'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /users/{userId}/roles:
    parameters:
      - $ref: '#/components/parameters/UserId'
    put:
      tags: [Users]
      summary: Replace the roles assigned to a user
      operationId: setUserRoles
      x-required-roles: [ADMIN]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserRolesRequest'
      responses:
        '200':
          description: Roles updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserProfile'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '429':
          $ref: '#/components/responses/TooManyRequests'
        '500':
          $ref: '#/components/responses/InternalServerError'
...
  /advisors:
    get:
      tags: [Advisors]
      summary: List advisors
      operationId: listAdvisors
      x-required-roles: [ADMIN, AUDITOR, ANALYST]
      responses:
        '200':
          description: Advisor list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdvisorListResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '429':
          $ref: '#/components/responses/TooManyRequests'
        '500':
          $ref: '#/components/responses/InternalServerError'
    post:
      tags: [Advisors]
      summary: Create an advisor record
      operationId: createAdvisor
      x-required-roles: [ADMIN]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateAdvisorRequest'
      responses:
        '201':
          description: Advisor created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdvisorSummary'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '409':
          $ref: '#/components/responses/Conflict'
        '429':
          $ref: '#/components/responses/TooManyRequests'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /advisors/{advisorId}:
    parameters:
      - $ref: '#/components/parameters/AdvisorId'
    get:
      tags: [Advisors]
      summary: Get advisor details
      operationId: getAdvisorById
      x-required-roles: [ADMIN, AUDITOR, ANALYST, ADVISOR]
      responses:
        '200':
          description: Advisor record
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdvisorSummary'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '429':
          $ref: '#/components/responses/TooManyRequests'
        '500':
          $ref: '#/components/responses/InternalServerError'
    patch:
      tags: [Advisors]
      summary: Update advisor metadata
      operationId: updateAdvisor
      x-required-roles: [ADMIN]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateAdvisorRequest'
      responses:
        '200':
          description: Advisor updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdvisorSummary'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '429':
          $ref: '#/components/responses/TooManyRequests'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /advisors/{advisorId}/clients:
    parameters:
      - $ref: '#/components/parameters/AdvisorId'
    get:
      tags: [Advisors]
      summary: List clients assigned to an advisor
      operationId: listAdvisorClients
      x-required-roles: [ADMIN, ADVISOR]
      responses:
        '200':
          description: Clients for advisor
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ClientListResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '429':
          $ref: '#/components/responses/TooManyRequests'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /clients:
    get:
      tags: [Clients]
      summary: List clients
      operationId: listClients
      x-required-roles: [ADMIN, AUDITOR, ANALYST, ADVISOR]
      responses:
        '200':
          description: Client list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ClientListResponse'
    post:
      tags: [Clients]
      summary: Create a client record
      operationId: createClient
      x-required-roles: [ADMIN, ADVISOR]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateClientRequest'
      responses:
        '201':
          description: Client created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ClientSummary'
  /clients/{clientId}:
    parameters:
      - $ref: '#/components/parameters/ClientId'
    get:
      tags: [Clients]
      summary: Get client details
      operationId: getClientById
      x-required-roles: [ADMIN, AUDITOR, ANALYST, ADVISOR, CLIENT]
      responses:
        '200':
          description: Client summary
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ClientSummary'
    patch:
      tags: [Clients]
      summary: Update a client record
      operationId: updateClient
      x-required-roles: [ADMIN, ADVISOR]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateClientRequest'
      responses:
        '200':
          description: Client updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ClientSummary'
  /clients/{clientId}/balance:
    parameters:
      - $ref: '#/components/parameters/ClientId'
    get:
      tags: [Clients]
      summary: Get the client cash balance
      operationId: getClientBalance
      x-required-roles: [ADMIN, AUDITOR, ANALYST, ADVISOR, CLIENT]
      responses:
        '200':
          description: Client cash balance
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ClientBalanceResponse'
...
  /analytics/advisors/{advisorId}:
    parameters:
      - $ref: '#/components/parameters/AdvisorId'
    get:
      tags: [Analytics]
      summary: Return analytics for a specific advisor
      operationId: getAdvisorAnalytics
      x-required-roles: [ADMIN, ANALYST, ADVISOR]
      responses:
        '200':
          description: Advisor analytics
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AnalyticsSummaryResponse'
  /analytics/clients/{clientId}:
    parameters:
      - $ref: '#/components/parameters/ClientId'
    get:
      tags: [Analytics]
      summary: Return analytics for a specific client
      operationId: getClientAnalytics
      x-required-roles: [ADMIN, ANALYST, ADVISOR, CLIENT]
      responses:
        '200':
          description: Client analytics
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AnalyticsSummaryResponse'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    UserId:
      name: userId
      in: path
      required: true
      schema:
        type: integer
        format: int64
...
    AdvisorId:
      name: advisorId
      in: path
      required: true
      schema:
        type: integer
        format: int64
    ClientId:
      name: clientId
      in: path
      required: true
      schema:
        type: integer
        format: int64
...
  responses:
    BadRequest:
      description: Invalid or malformed request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiErrorResponse'
    Unauthorized:
      description: Missing or invalid authentication
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiErrorResponse'
    Forbidden:
      description: Authenticated user does not have permission to access the resource
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiErrorResponse'
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiErrorResponse'
    Conflict:
      description: Duplicate resource or conflicting state
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiErrorResponse'
    TooManyRequests:
      description: Too many requests were submitted in a short time window
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiErrorResponse'
    InternalServerError:
      description: Unexpected server-side failure
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiErrorResponse'
  schemas:
    ApiErrorResponse:
      type: object
      required: [timestamp, status, error, message, path]
      properties:
        timestamp:
          type: string
          format: date-time
        status:
          type: integer
          format: int32
        error:
          type: string
        message:
          type: string
        path:
          type: string
...
    UserProfile:
      type: object
      properties:
        userId:
          type: integer
          format: int64
        username:
          type: string
        email:
          type: string
          format: email
        displayName:
          type: string
        enabled:
          type: boolean
        roles:
          type: array
          items:
            type: string
        advisorId:
          type: integer
          format: int64
          nullable: true
        clientId:
          type: integer
          format: int64
          nullable: true
...
    UserListResponse:
      type: object
      properties:
        users:
          type: array
          items:
            $ref: '#/components/schemas/UserProfile'
    UpdateUserRequest:
      type: object
      properties:
        displayName:
          type: string
        email:
          type: string
          format: email
        enabled:
          type: boolean
    UserRolesRequest:
      type: object
      required: [roles]
      properties:
        roles:
          type: array
          items:
            type: string
    AdvisorSummary:
      type: object
      properties:
        advisorId:
          type: integer
          format: int64
        advisorName:
          type: string
        userId:
          type: integer
          format: int64
          nullable: true
    AdvisorListResponse:
      type: object
      properties:
        advisors:
          type: array
          items:
            $ref: '#/components/schemas/AdvisorSummary'
    CreateAdvisorRequest:
      type: object
      required: [advisorName]
      properties:
        advisorName:
          type: string
        userId:
          type: integer
          format: int64
          nullable: true
    UpdateAdvisorRequest:
      type: object
      properties:
        advisorName:
          type: string
        userId:
          type: integer
          format: int64
          nullable: true
    ClientSummary:
      type: object
      properties:
        clientId:
          type: integer
          format: int64
        clientName:
          type: string
        advisorId:
          type: integer
          format: int64
          nullable: true
        modelPortfolioId:
          type: integer
          format: int64
          nullable: true
        cashBalance:
          type: number
          format: double
    ClientListResponse:
      type: object
      properties:
        clients:
          type: array
          items:
            $ref: '#/components/schemas/ClientSummary'
    CreateClientRequest:
      type: object
      required: [clientName]
      properties:
        clientName:
          type: string
        advisorId:
          type: integer
          format: int64
          nullable: true
        modelPortfolioId:
          type: integer
          format: int64
          nullable: true
        createdByUserId:
          type: integer
          format: int64
          nullable: true
        cashBalance:
          type: number
          format: double
    UpdateClientRequest:
      type: object
      properties:
        clientName:
          type: string
        advisorId:
          type: integer
          format: int64
          nullable: true
        modelPortfolioId:
          type: integer
          format: int64
          nullable: true
        cashBalance:
          type: number
          format: double
    ClientBalanceResponse:
      type: object
      properties:
        clientId:
          type: integer
          format: int64
        cashBalance:
          type: number
          format: double
...
    AnalyticsSummaryResponse:
      type: object
      properties:
        scope:
          type: string
          description: Scope of the analytics payload, such as admin, advisor, client, or portfolio
        generatedAt:
          type: string
          format: date-time
        metrics:
          type: object
          additionalProperties: true
        series:
          type: array
          items:
            type: object
            additionalProperties: true
...